+++ src/main/java/com/taw/polybank/dao/TransactionRepository.java
@@ -15,16 +15,16 @@ public interface TransactionRepository extends JpaRepository<TransactionEntity, 
     public List<TransactionEntity> findByBankAccountByBankAccountId(BankAccountEntity bankAccount);
 
     @Query("select t from TransactionEntity t where t.bankAccountByBankAccountId = :bankAccount and t.timestamp >= :begin and t.timestamp <= :end and t.clientByClientId in :clients and t.paymentByPaymentId.benficiaryByBenficiaryId = :beneficiary and t.paymentByPaymentId.amount >= :amount")
-    List<TransactionEntity> filterByBankAccount_TimestampRange_TransactionOwner_beneficiaryIBAN_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, List<ClientEntity> clients, BenficiaryEntity beneficiary, int amount);
+    List<TransactionEntity> filterByBankAccount_TimestampRange_TransactionOwner_beneficiaryIBAN_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, List<ClientEntity> clients, BenficiaryEntity beneficiary, double amount);
 
     @Query("select t from TransactionEntity t where t.bankAccountByBankAccountId = :bankAccount and t.timestamp >= :begin and t.timestamp <= :end and t.paymentByPaymentId.benficiaryByBenficiaryId = :beneficiary and t.paymentByPaymentId.amount >= :amount")
-    List<TransactionEntity> filterByBankAccount_TimestampRange_beneficiaryIBAN_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, BenficiaryEntity beneficiary, int amount);
+    List<TransactionEntity> filterByBankAccount_TimestampRange_beneficiaryIBAN_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, BenficiaryEntity beneficiary, double amount);
 
     @Query("select t from TransactionEntity t where t.bankAccountByBankAccountId = :bankAccount and t.timestamp >= :begin and t.timestamp <= :end and t.clientByClientId in :clients and t.paymentByPaymentId.amount >= :amount")
-    List<TransactionEntity> filterByBankAccount_TimestampRange_TransactionOwner_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, List<ClientEntity> clients, int amount);
+    List<TransactionEntity> filterByBankAccount_TimestampRange_TransactionOwner_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, List<ClientEntity> clients, double amount);
 
     @Query("select t from TransactionEntity t where t.bankAccountByBankAccountId = :bankAccount and t.timestamp >= :begin and t.timestamp <= :end and t.paymentByPaymentId.amount >= :amount")
-    List<TransactionEntity> filterByBankAccount_TimestampRange_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, int amount);
+    List<TransactionEntity> filterByBankAccount_TimestampRange_Amount(BankAccountEntity bankAccount, Timestamp begin, Timestamp end, double amount);
